refactor: extract number-field logic to reusable NumberFieldMixin
